Got tired of the mess of a whole file of Product classes, split it into a package.
The package has been more sensibly structured, with product types split into photometric, spectroscopic, miscellaneous, and base products.

Also added BaseAggregateProduct into the base.py file, though it doesn't do anything yet. It will be used for when many Products need to be brought together into one, such as PSF grids and annular spectra.
